fix: move openapi inline test to sibling _tests.rs file (#64)
The pre-push hook (.githooks/pre-push) enforces a test-file convention:
tests must live in `*_tests.rs` files, not inline `#[cfg(test)] mod foo {}`
blocks. `src/openapi.rs` was the sole violator — its inline `mod tests`
block tripped the convention check and failed the hook on every push.

Move the `committed_spec_is_current` test into a new `src/openapi_tests.rs`
sibling and declare it with `#[path = "openapi_tests.rs"] mod openapi_tests;`,
matching the pattern already used by `error.rs`, `time.rs`, etc. No behavior
change — the test still runs and passes.
